[BOJ]15666.N과 M(7)
