+++ src/boj/15656/README.md
@@ -1,0 +1,37 @@
+# [Silver III] N과 M (7) - 15656
+
+[문제 링크](https://www.acmicpc.net/problem/15656)
+
+### 성능 요약
+
+메모리: 157340 KB, 시간: 692 ms
+
+### 분류
+
+백트래킹
+
+### 제출 일자
+
+2024년 8월 13일 03:40:26
+
+### 문제 설명
+
+<p>N개의 자연수와 자연수 M이 주어졌을 때, 아래 조건을 만족하는 길이가 M인 수열을 모두 구하는 프로그램을 작성하시오. N개의 자연수는 모두 다른 수이다.</p>
+
+<ul>
+	<li>N개의 자연수 중에서 M개를 고른 수열</li>
+	<li>같은 수를 여러 번 골라도 된다.</li>
+</ul>
+
+### 입력
+
+ <p>첫째 줄에 N과 M이 주어진다. (1 ≤ M ≤ N ≤ 7)</p>
+
+<p>둘째 줄에 N개의 수가 주어진다. 입력으로 주어지는 수는 10,000보다 작거나 같은 자연수이다.</p>
+
+### 출력
+
+ <p>한 줄에 하나씩 문제의 조건을 만족하는 수열을 출력한다. 중복되는 수열을 여러 번 출력하면 안되며, 각 수열은 공백으로 구분해서 출력해야 한다.</p>
+
+<p>수열은 사전 순으로 증가하는 순서로 출력해야 한다.</p>
+
